refactor sqlalchemy models
